Fixed name assignment on RteSolver object
Initializing a generic RteSolver() object without scene parameter args (as in the optimize extinction script) fails as _wavelen is not yet assigned until scene parameters are loaded so name assignment fails. 

Moved the default name assignment inside set_scene() so that _wavelen is known.
If there are no scene parameters assigned name is left as None.